arm soft hard object mix test

+++ c/arm_softfp_hard_mix_test/megacrap.h
@@ -1,7 +1,7 @@
 #ifndef MEGACRAP_H
 #define MEGACRAP_H
 
-	int megacrap(int);
+	float megacrap(int);
 
 #endif
